move SimLaserScanner to its own class in the gl_depth_sim library (where SimDepthCamera and other classes are stored). Does not allow for ROS or PCL, so implemented similar to SimDepthCamera where return DepthImage that gets converted to a pointcloud. See prior commit to have SimLaserScanner::render return a pointcloud
